+++ Next.PCL/Entities/Metas/MetaVideo.cs
@@ -8,9 +8,9 @@ namespace Next.PCL.Entities
     {
         public Uri Url { get; set; }
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public int Width { get; set; }
+        public ushort Width { get; set; }
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public int Height { get; set; }
+        public ushort Height { get; set; }
         public MetaSource Source { get; set; }
         public MetaVideoType Type { get; set; }
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
